Tasks until 9
First 5 tasks finished.
Task 6: there are problems with Noise Bilateral. We don't use v_ij and
the result we think is not the correct.
Task 7: It's OK, but in Openin and Closing, the value of m_type always
is Open.
Task 8: There are an error in Roberts, Prewitt and Sobel because of
array out of bounds in transformation.cpp. If we change the loop and we
do one iteraction less (we change <= by <), it runs, but then doesn't
run the Blurring.
In the Zero-Crossing always appear a black image. The problem can be
because we don't calculate well the Laplacian of Gauss of the input
image.